requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
bleach==4.1.0
certifi==2021.10.8
cycler==0.11.0
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-admin-list-filter-dropdown==1.0.3
django-countries==7.2.1
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-quill-editor==0.1.10
django-summernote==0.8.20.0
django-tinymce==3.4.0
django-widget-tweaks==1.4.12
fonttools==4.29.1
greenlet==1.1.2
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.0
psycopg2 @ file:///C:/ci/psycopg2_1612298777194/work
pydotplus==2.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
webencodings==0.5.1
wincertstore==0.2
